// src/app/shared/enums/Beach.ts
export enum SantosBeachEnum {
  JOSE_MENINO = 'JOSE_MENINO',
  GONZAGA = 'GONZAGA',
  BOQUEIRAO = 'BOQUEIRAO',
  EMBARE = 'EMBARE',
  APARECIDA = 'APARECIDA',
  PONTA_DA_PRAIA = 'PONTA_DA_PRAIA',
  POMPEIA = 'POMPEIA'
}

export enum SaoVicenteBeachEnum {
  ITARARE = 'ITARARE',
  MILIONARIOS = 'MILIONARIOS',
  GONZAGUINHA = 'GONZAGUINHA',
  PARANAPUA = 'PARANAPUA',
  ITAQUITANDUVA = 'ITAQUITANDUVA'
}